Fix for adding/deleting Recipe images/cards

+++ ffreact-prod/src/Recipe/Recipe.js
@@ -1,4 +1,4 @@
-import { Button, Typography, Box, Grid, Checkbox, Snackbar, FormControlLabel, TextField, InputLabel, Paper } from "@mui/material";
+import { Button, Typography, Box, Grid, Checkbox, Snackbar, FormControlLabel, TextField, InputLabel, Paper, Popover } from "@mui/material";
 import { HighlightOff } from "@mui/icons-material";
 import React, {useState, useEffect, useCallback, Fragment, useRef} from 'react';
 import { DataGrid, useGridApiContext } from "@mui/x-data-grid";
@@ -145,35 +145,46 @@ export default function Recipe({recipeData, setRecipeData, ingredientOptions, pa
     const [imageFile, setImageFile] = useState();
     const [cardFile, setCardFile] = useState();
     
-    const [imageURL, setImageURL] = useState();
-    const [cardURL, setCardURL] = useState();
+    const [imageURL, setImageURL] = useState(recipeData ? recipeData.r_img_path : undefined);
+    const [cardURL, setCardURL] = useState(recipeData ? recipeData.r_card_path : undefined);
+
+    const [deleteImage, setDeleteImage] = useState(false);
+    const [deleteCard, setDeleteCard] = useState(false);
+
+    const [popoverAnchors, setPopoverAnchors] = useState({confirmDeleteImageAnchor: null, confirmDeleteCardAnchor: null});
+    const [confirmDeleteImageOpen, setConfirmDeleteImageOpen] = useState(false);
+    const [confirmDeleteCardOpen, setConfirmDeleteCardOpen] = useState(false);
 
     const handleCloseClick = () => {
         // Return to recipe list when close is clicked
-        axios({
-            method: "DELETE",
-            url:"http://4.236.185.213:8000/api/" + 'tempimageupload' + '/',
-            data: imageURL
-        }).then((response)=>{
-        }).catch((error) => {
-        if (error.response) {
-            console.log(error.response);
-            console.log(error.response.status);
-            console.log(error.response.headers);
-            }
-        });
-        axios({
-            method: "DELETE",
-            url:"http://4.236.185.213:8000/api/" + 'tempcardupload' + '/',
-            data: cardURL
-        }).then((response)=>{
-        }).catch((error) => {
-        if (error.response) {
-            console.log(error.response);
-            console.log(error.response.status);
-            console.log(error.response.headers);
-            }
-        });
+        if (imageURL) {
+            axios({
+                method: "DELETE",
+                url:"http://4.236.185.213:8000/api/" + 'tempimageupload' + '/',
+                data: imageURL
+            }).then((response)=>{
+            }).catch((error) => {
+            if (error.response) {
+                console.log(error.response);
+                console.log(error.response.status);
+                console.log(error.response.headers);
+                }
+            });
+        }
+        if (cardURL) {
+            axios({
+                method: "DELETE",
+                url:"http://4.236.185.213:8000/api/" + 'tempcardupload' + '/',
+                data: cardURL
+            }).then((response)=>{
+            }).catch((error) => {
+            if (error.response) {
+                console.log(error.response);
+                console.log(error.response.status);
+                console.log(error.response.headers);
+                }
+            });
+        }
         setCurrPage(<RecipePage setCurrPage={setCurrPage}></RecipePage>)
     }
 
@@ -202,7 +213,53 @@ export default function Recipe({recipeData, setRecipeData, ingredientOptions, pa
         });
     }
 
+    const handleDeleteRecipeImage = (imgOrCard) => {
+        if (recipeData.r_img_path || recipeData.r_card_path) {
+            axios({
+                method: "DELETE",
+                url:"http://4.236.185.213:8000/api/" + (imgOrCard==='image' ? 'mealrecipe-image' : 'mealrecipe-card') + "/" + recipeData.r_num
+            }).then((response)=>{
+            }).catch((error) => {
+            if (error.response) {
+                console.log(error.response);
+                console.log(error.response.status);
+                console.log(error.response.headers);
+                }
+            });
+        }
+    }
+
+    const handleDeleteTempImage = (imgOrCard) => {
+        if (imageURL && !recipeData.r_img_path || cardURL && !recipeData.r_card_path) {
+            axios({
+                method: "DELETE",
+                url:"http://4.236.185.213:8000/api/" + (imgOrCard==='image' ? 'tempimageupload' : 'tempcardupload') + '/',
+                data: imgOrCard==='image' ? imageURL : cardURL
+            }).then((response)=>{
+            }).catch((error) => {
+            if (error.response) {
+                console.log(error.response);
+                console.log(error.response.status);
+                console.log(error.response.headers);
+                }
+            });
+        }
+        if (imgOrCard==='image') {
+            setImageFile();
+            setImageURL();
+            setDeleteImage(true);
+        }
+        else if (imgOrCard==='card') {
+            setCardFile();
+            setCardURL();
+            setDeleteCard(true);
+        }
+    }
+
     const handleImageUpload = (file, r_num, apiEndpoint) => {
+        if (!file) {
+            return;
+        }
         // Send file in request to api
         const formData = new FormData();
         formData.append('file', file);
@@ -240,10 +297,18 @@ export default function Recipe({recipeData, setRecipeData, ingredientOptions, pa
         const r_data = {...recipeData, r_name: nameField.current.value, r_ingredients: ingredientRows, r_packaging: packagingRows, r_stations: stationRows, m_s: m_s}
         // console.log(JSON.stringify(r_data));
         setUpdateSBOpen(true);
+
+        if (deleteImage) {
+            handleDeleteRecipeImage('image')
+        }
+        if (deleteCard) {
+            handleDeleteRecipeImage('card')
+        }
+
         if (isAdding) {
             axios({
                 method: "POST",
-                url:"http://localhost:8000/api/mealrecipes/",
+                url:"http://4.236.185.213:8000/api/mealrecipes/",
                 data: r_data,
             }).then((response)=>{
                 handleImageUpload(imageFile, response.data, 'mealrecipe-image');
@@ -262,7 +327,7 @@ export default function Recipe({recipeData, setRecipeData, ingredientOptions, pa
         else {
             axios({
                 method: "PATCH",
-                url:"http://localhost:8000/api/mealrecipes/" + recipeData.r_num + '/',
+                url:"http://4.236.185.213:8000/api/mealrecipes/" + recipeData.r_num + '/',
                 data: r_data,
             }).then((response)=>{
                 handleImageUpload(imageFile, recipeData.r_num, 'mealrecipe-image');
@@ -294,7 +359,7 @@ export default function Recipe({recipeData, setRecipeData, ingredientOptions, pa
         if (props.image_source) {
             return (
                 <Box sx={{position: 'relative'}}>
-                    <Button color='lightBlue' variant='contained' sx={{position: 'absolute', right: '0%'}} onClick={() => console.log(imageFile)}>
+                    <Button color='lightBlue' variant='contained' sx={{position: 'absolute', right: '0%'}} onClick={() => handleDeleteTempImage('image')}>
                         <HighlightOff/>
                     </Button>
                     <Box sx={{height: '100%', width: '100%'}}>
@@ -314,7 +379,7 @@ export default function Recipe({recipeData, setRecipeData, ingredientOptions, pa
         if (props.card_source) {
             return (
                 <Box sx={{position: 'relative'}}>
-                    <Button color='lightBlue' variant='contained' sx={{position: 'absolute', right: '0%'}} onClick={() => console.log(imageFile)}>
+                    <Button color='lightBlue' variant='contained' sx={{position: 'absolute', right: '0%'}} onClick={() => {handleDeleteTempImage('card')}}>
                         <HighlightOff/>
                     </Button>
                     <Box sx={{height: '100%', width: '100%'}}>
@@ -356,14 +421,14 @@ export default function Recipe({recipeData, setRecipeData, ingredientOptions, pa
                     <FormControlLabel control={<Checkbox checked={m_s===1 ? true : false} onChange={handleMealSnackChange}/>} label="Meal Recipe?"/>
                     
                     {/* Recipe Image */}
-                    <RecipeImage image_source={imageURL ? imageURL : recipeData.r_img_path}/>
+                    <RecipeImage image_source={imageURL}/>
                     <Button color='lightBlue' variant='contained' component='label'>
                         Upload Image
                         <input id='recipe_image' type='file' accept='.jpg,.png,.bmp' onChange={(event) => {handleTempUpload(event.target.files[0], 'tempimageupload'); setImageFile(event.target.files[0])}} hidden></input>
                     </Button>
 
                     {/* Recipe Card */}
-                    <RecipeCard card_source={cardURL ? cardURL : recipeData.r_card_path}/>
+                    <RecipeCard card_source={cardURL}/>
                     <Button color='lightBlue' variant='contained' component='label'>
                         Upload Recipe Card
                         <input id='recipe_card' type='file' accept='.jpg,.pdf,.doc,.docx' onChange={(event) => {handleTempUpload(event.target.files[0], 'tempcardupload'); setCardFile(event.target.files[0])}} hidden></input>
